Ensure the queue is large enough
The maximum size of the queue is the number of unique friends in all pairings.
Thus, the size of the ringBuffer needs to be able to accomodate `2n` elements.
For example, consider the case where the input is `[[1, 2], [3, 4], [5, 6], [7, 8]]`,
`n` is equal to 4 because there are 4 pairings, but there are 8 friends.

Signed-off-by: Peter Engelbert <pmengelbert@gmail.com>